fix: move workspace types to types files and use notes form directly
