+++ packages/esm-ward-app/src/index.ts
@@ -8,7 +8,6 @@ import {
 import { configSchema } from './config-schema';
 import rootComponent from './root.component';
 import { moduleName } from './constant';
-import WardPatientActionButton from './ward-patient-workspace/ward-patient-action-button.extension';
 import { createDashboardLink } from './createDashboardLink.component';
 
 export const importTranslation = require.context('../translations', false, /.json$/, 'lazy');
@@ -39,19 +38,19 @@ export const wardPatientWorkspace = getAsyncLifecycle(
   options,
 );
 
+// t("inpatientNotesWorkspaceTitle", "In-patient notes")
 export const wardPatientNotesWorkspace = getAsyncLifecycle(
   () => import('./ward-workspace/ward-patient-notes/notes.workspace'),
   options,
 );
 
-export const wardPatientActionButtonExtension = getSyncLifecycle(WardPatientActionButton, options);
+export const wardPatientActionButtonExtension = getAsyncLifecycle(
+  () => import('./ward-patient-workspace/ward-patient-action-button.extension'),
+  options,
+);
 
 export const wardPatientNotesActionButtonExtension = getAsyncLifecycle(
   () => import('./ward-workspace/ward-patient-notes/notes-action-button.extension'),
-  options,
-);
-export const wardPatientNotesFormExtension = getAsyncLifecycle(
-  () => import('./ward-workspace/ward-patient-notes/form/notes-form.component'),
   options,
 );
