+++ packages/esm-ward-app/src/ward-patient-workspace/types.ts
@@ -1,0 +1,6 @@
+import type { DefaultWorkspaceProps, PatientUuid } from '@openmrs/esm-framework';
+import type { WardPatientCardProps } from '../types';
+
+export interface WardPatientWorkspaceProps extends DefaultWorkspaceProps, WardPatientCardProps {
+  patientUuid: PatientUuid;
+}
