+++ packages/esm-ward-app/src/ward-workspace/ward-patient-notes/types.ts
@@ -1,0 +1,4 @@
+import type { DefaultWorkspaceProps } from '@openmrs/esm-framework';
+import { type WardPatientCardProps } from '../../types';
+
+export interface WardPatientNotesWorkspaceProps extends DefaultWorkspaceProps, WardPatientCardProps {}
