+++ packages/esm-ward-app/translations/en.json
@@ -19,28 +19,29 @@
   "failedToLoadPatientWorkspace": "Ward patient workspace has failed to load.",
   "female": "Female",
   "fetchingEmrConfigurationFailed": "Fetching EMR configuration failed. Try refreshing the page or contact your system administrator.",
+  "inpatientNotesWorkspaceTitle": "In-patient notes",
   "invalidLocationSpecified": "Invalid location specified",
   "invalidWardLocation": "Invalid ward location: {{location}}",
   "male": "Male",
   "manage": "Manage",
   "noBedsConfigured": "No beds configured for this location",
   "noBedsConfiguredForLocation": "No beds configured for {{location}} location",
+  "note": "Note",
   "other": "Other",
   "patientAdmittedButBedNotAssigned": "Patient admitted successfully but fail to assign bed to patient",
   "patientAdmittedSuccessfully": "Patient admitted successfully",
   "patientAdmittedSuccessfullySubtitle": "{{patientName}} has been successfully admitted and assigned to bed {{bedNumber}}",
   "patientAdmittedWoBed": "Patient admitted successfully to {{location}}",
+  "patientNoteNowVisible": "It should be now visible in the notes history",
+  "patientNoteSaveError": "Error saving patient note",
+  "save": "Save",
+  "saving": "Saving...",
   "searchForPatient": "Search for a patient",
   "selectABed": "Select a bed",
   "somePartsOfTheFormDidntLoad": "Some parts of the form didn't load",
   "transferElsewhere": "Transfer elsewhere",
   "unableToSelectBeds": "Unable to select beds",
   "unknown": "Unknown",
-  "note": "Note",
-  "patientNoteNowVisible": "It should be now visible in the notes history",
-  "patientNoteSaveError": "Error saving patient note",
-  "save": "Save",
-  "saving": "Saving...",
   "visitNoteSaved": "Patient note saved",
   "wardClinicalNotePlaceholder": "Write any notes here",
   "wardPatientWorkspaceTitle": "Ward Patient",
